refactor(db): split in-memory trie module by scope
