+++ madara/crates/client/db/src/rocksdb/global_trie/in_memory/mod.rs
@@ -1,0 +1,14 @@
+mod compute;
+mod db;
+mod overlay;
+mod state_diff;
+
+#[cfg(test)]
+mod tests;
+
+pub use compute::{
+    compute_root_from_snapshot, compute_roots_in_parallel_from_snapshot, flush_overlay_and_checkpoint,
+    InMemoryRootComputation,
+};
+pub use overlay::{BonsaiOverlay, TrieLogMode};
+pub use state_diff::{cumulative_squashed_state_diffs, squash_state_diffs};
